% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the testapp}
\usage{
run_app(options = list())
}
\arguments{
\item{options}{Named options that should be passed to the run_app call.}
}
\value{
Open browser
}
\description{
This function is used to generate the standalone version
}
